refactor for custom hooks
