+++ src/components/ReviewPage.jsx
@@ -1,6 +1,5 @@
-import { useEffect, useState } from "react";
 import { Link, useParams } from "react-router-dom";
-import { fetchReviewById } from "../api";
+import { useReviewById } from "../hooks/useReviewById";
 import styles from "../modules/ReviewPage.module.css";
 import { ErrorHandling } from "./ErrorHandling";
 
@@ -15,25 +14,13 @@ const {
 } = styles;
 
 export const ReviewPage = () => {
-  const [isLoading, setIsLoading] = useState(true);
-  const [review, setReview] = useState({});
-  const [error, setError] = useState(null);
-
   const { review_id } = useParams();
+  const { error, review, isLoading } = useReviewById(review_id);
   const { category, review_img_url, designer, votes, title, owner } = review;
-
-  useEffect(() => {
-    fetchReviewById(review_id)
-      .then(({ review }) => {
-        setIsLoading(true);
-        setReview(review);
-        setIsLoading(false);
-      })
-      .catch((err) => setError(err));
-  }, [review_id]);
 
   if (error) return <ErrorHandling error={error} />;
   if (isLoading) return <p>Loading...</p>;
+
   return (
     <main>
       <section className={reviewWrapper}>
